[PyTorch][Dispatcher] Fix destruction order fiasco crash (#104393)
Summary:
The current implementation of `Dispatcher` returns an RAII object
from it's `register*` methods which, on destruction, uses a saved
reference to the `Dispatcher` to call the associated `deregister*`
method.

However, nothing guarantees that the `Dispatcher` is destroyed
*after* all RAII objects have been destroyed and, in practice, we
see segfaults caused when a global `Dispatcher` is cleaned up
before RAII globals.

This diff fixes by keeping the `Dispatcher` lock and "alive" marker
in a `std::shared_ptr` which the callbacks copy and then use to
verify that the `Dispatcher` is still alive before continuing.

https://fb.workplace.com/groups/1405155842844877/posts/7143161099044294/
https://fb.workplace.com/groups/python.builds/posts/3510588832595867/
S349108

Test Plan: CI

Differential Revision: D47113122

Pull Request resolved: https://github.com/pytorch/pytorch/pull/104393
Approved by: https://github.com/ezyang