[inductor] Fix xpasses being impossible (#106631)
This test raises an error inside the test when an xfailed test succeeds, but
is also decorated with the xfail decorator which converts the error to an xfail.
Instead, this lets the test function pass normally and lets the xfail decorator
raise "Unexpected success".

I also updated the COLLECT_EXPECT code and run it to get the updated set of
failures.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/106631
Approved by: https://github.com/lezcano
ghstack dependencies: #106319, #106400